Google Play apps to unmanaged devices
It is not straight forward to link the latest changes introduced to **Device Enrollment setting** option introduced in Company Portal customization to the scenario of deploying apps to unmanaged devices. I added a couple of lines that might help understand how this feature works in this scenario. 
Based on customer feedback, they are not able to come to the conclusion that that **Device Enrollment setting** will help prevent users from enrolling their device unintentionally or be confused with notifications asking them to enroll their device when it is not the company's intention.